requirements.txt
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-paypal==2.0

